chatbot-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2

# Gemini SDKs
google-genai==1.0.0
google-generativeai

# RAG
chromadb==0.5.5
sentence-transformers==3.0.1
tiktoken==0.7.0

# File parsing
pypdf==4.3.1
python-docx==1.1.2
python-pptx==0.6.23
pandas==2.2.2
numpy==1.26.4
Pillow==10.4.0

# Voice + Vision (Whisper + BLIP/CLIP)
faster-whisper==1.0.3
torch==2.3.1
transformers==4.43.4
opencv-python==4.10.0.84
git+https://github.com/openai/CLIP.git

# Environment & helpers
python-dotenv==1.0.1
tqdm

# Database + ORM
sqlalchemy

# Web scraping
requests
beautifulsoup4
lxml

# SQL Server driver
pyodbc

# Auth hashing
passlib[bcrypt]
